+++ provider_verifier/provider_verifier_download.go
@@ -1,4 +1,4 @@
-package provider_verification
+package provider_verifier
 
 import (
 	"context"
